% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePanel.R
\name{EvaluatePanel}
\alias{EvaluatePanel}
\title{Evaluate a panel with systemic effectiveness indicators}
\usage{
EvaluatePanel(strpath, raremode = "ISFG", Nind = "lastrow")
}
\arguments{
\item{strpath}{name of .csv file containing af matrix, with the marker names listed in the first row}

\item{raremode}{mode of the calculation method of rare alleles,with a default of "ISFG", indicating the recommended method from ISFG, i.e., (X+1)/(2N+1), where X and N stood for the number of allele types detected in a survey and the sample size, respectively. Two alternative methods are given: "MAF", take the minimum allele frequency as such frequency; and "1/2N", take the minimum of possible frequency a survey can achieave.}

\item{Nind}{mode of sample size, with a default of "lastrow", meaning that the sample size was presented in the last row of the .csv file. An alternative method is given, i.e., input a unified sample size.}
}
\value{
list of four vectors: afmatrix, a list of allele frequency data of each locus; rare, a data.frame containing the frequency of rare allele on each locus; indicators, a data.frame containing parameters of system efficiency for each locus; panelparas, a data.frame containing system efficiency parameters for the whole panel, with the form of log10(1-paramter) to avoid the situation that the parameters being displayed as 1 because they were too close to 1
}
\description{
Evaluate a panel with systemic effectiveness indicators
}
