% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FortytwoSTR}
\alias{FortytwoSTR}
\title{Information of 42 autosomal STRs}
\format{
A list of 3 vectors: i) a list nameed afmatrix containing 42 data.frames, each of which exhibiting the allele frequecies of alleles on a locus with the row names being the corresponding names of such alleles; ii) a data.fram with 1 row and 42 columns containing the frequencies of rare alleles on the 42 loci; and iii) a data.frame named indicators with 42 columns and 12 rows containing the main effeciency indicators of the 42 loci.
}
\usage{
FortytwoSTR
}
\description{
Example of the results achieved by "EvaluatePanel" function
}
\keyword{datasets}
