% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IICAL.R
\name{IICAL}
\alias{IICAL}
\title{calculation of Incest index}
\usage{
IICAL(Parent, Child, af, rare, theta)
}
\arguments{
\item{Parent}{Genotypes of individual A of each case, which should be data.frame with 2 columns and ss rows, where ss stand for sample size}

\item{Child}{Genotypes of individual B of each case, which should be data.frame with 2 columns and ss rows, where ss stand for sample size}

\item{af}{name of allele frequency matrix, a data.frame of 1 column with the allele name being row names}

\item{rare}{frequency of rare allele}

\item{theta}{kinship coefficent between the parents under Hp (that under Hd euquals to 0)}
}
\value{
a data.frame containing 3 columns: relationship, IBS and PI for each simulation
}
\description{
Calculation of the ratio for a parent-child pair between the likelihood that the child's father is a relative of the parent to that is unrelated
}
