% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KINSIMU-package.R
\docType{package}
\name{KINSIMU-package}
\alias{KINSIMU}
\alias{KINSIMU-package}
\title{KINSIMU: Panel Evaluation in Forensic Kinship Analysis}
\description{
Evaluate specific panels in different aspects: i) Simulation tools related to pedigree researches; ii) calculation for systemic effectiveness indicators, such as probability of exclusion (PE).
}
\author{
\strong{Maintainer}: Guanju Ma \email{mache124@126.com}

Other contributors:
\itemize{
  \item Shujin Li [contributor]
}

}
\keyword{internal}
