% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRgpgcam.R
\name{LRgpgcam}
\alias{LRgpgcam}
\title{LR in grandparent-child identification with the reference of a full-sibling of the alleged parent between them}
\usage{
LRgpgcam(A, B, GP, af, rare = NULL, allelename = FALSE, P = NULL)
}
\arguments{
\item{A}{Genotype data of the alleged uncle/aunt, should be data.frame with 2 columns and ss rows, where ss stand for sample size;}

\item{B}{Genotype data of the child, with the same form with \code{A}}

\item{GP}{Genotype data of the alleged grandparent of B, with the same form with \code{A}}

\item{af}{name of allele frequency matrix, a data.frame of 1 column containing frequencies with allele names being row names}

\item{rare}{frequency of rare allele on the locus}

\item{allelename}{if TRUE, the input genotype data would be regarded as allelenames, otherwise, the position in the af matrix}

\item{P}{Genotype data of the other parent of B, which can be \code{NULL} or as that of \code{A}}
}
\value{
a data frame with one column and ss rows, containing log10 value of the CLR of each case
}
\description{
LR when a child (B) is alleged to be a grand-child of another individual (GP), and an offspring(A) of the alleged grandparent participated, with or without the assistant of B's other parent (M). Hp is that, B is an offspring of A's full-sibling; and Hd that, B is unrelated to GP and A. Inbreeding is not considered.
}
\details{
There might be no allele sharing between GP and A, or between M and B, if so, the part in LR would be output as 0, which can be further optimized in future version.
}
