% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputCSV.R
\name{outputCSV}
\alias{outputCSV}
\title{Output frequency data into ISFG format}
\usage{
outputCSV(data, strpath)
}
\arguments{
\item{data}{frequency data, a list containing 4 data frames.}

\item{strpath}{name of the pathway of the output csv file}
}
\value{
a .csv file in ISFG formate
}
\description{
Output frequency data into ISFG format
}
\examples{
path<-tempdir()
outputCSV(FortytwoSTR,file.path(path,'data.csv'))
}
