% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{pediexample}
\alias{pediexample}
\title{Example of pedi matrix}
\format{
a data.frame containg 3 columns
}
\usage{
pediexample
}
\description{
Example of pedi, a matrix containing inherence information of a pedigree, used in function of "pedisimu"
}
\keyword{datasets}
