% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptoc.R
\name{ptoc}
\alias{ptoc}
\title{Parent-child inherence}
\usage{
ptoc(Parent = NULL, af = NULL, ss = NULL, mu = 0, allelename = FALSE)
}
\arguments{
\item{Parent}{information of Parents, with a default of NULL, standing for random male, otherwise it should be the name of a existing data.frame containing the Parents' genotypes, 2 columns × ss rows;}

\item{af}{name of allele frequency matrix, a data.frame of 1 column containing frequencies with allele names being row names, which can be loaded with "evaluate_panel" function, not necessary if Parent is not NULL}

\item{ss}{sample size, i.e., how many children do you want simulate, not necessary if Parent is not NULL}

\item{mu}{mutation rate, with a default of 0}

\item{allelename}{if TRUE the output data would be the allele name, which should be the row names of af matrix, otherwise, it would be the position in that matrix}
}
\value{
a list of two data.frame of 2 columns and ss rows containing genotypes of individual pairs with specific relationships
}
\description{
Generating alleles of multiple children on a single locus with one of their parents' genotypes
}
